input_buffer.c and Boss 4. (#129)
* Start sub_800DF8C

* Match most of sakit/input_buffer.c

* Very minor progress on sub_800DF8C

* Match Boss 4's bombs and debris

* Match TaskDestructor_AeroEggMain

* Match more Boss4 procedures

* Match sub_80424EC

* Match sub_80423EC

* Split AeroEgg struct into main/sub and match AeroEggResetPos

* Match sub_804217C

* Remove addresses

* Document some of AeroEggSub

* Add AE_UPDATE_PART_DEFEATED macro

* Rename to Player_DisableInputAndBossTimer/Player_DisableInputAndBossTimer_FinalBoss

* Get close to matching sub_8042024

* Check for boss lives properly

* sub_8042024 -> AeroEgg_InitPartsDefeated; move it into nonmatch

* Start sub_8041D34

* sub_8041D34 -> gTileInfoBossScrews

* Progress on sub_8041D34

* Progress on sub_8041D34

* Move sub_8041D34 into non-match

* Remove redundant struct copy

* Match sub_8041C48

* More Boss 4 procedures

* Get sub_8041A08 to ~91% and put into non-match

* Add Q/I aliases

* Add some comments to aliases, for consistency

* Match sub_8041A08

* Use Q/I macros

* Get sub_8041880 to 85.03% and move it into non-match

* Match Task_AeroEggExploding

* Match CreateAeroEgg. Remove boss_4.s

* Task_8042AB0 -> Task_AeroEggBombHitGround

* Progress on sub_800DF8C

* Progress on sub_800DF8C and move it into non-match. Remove input_buffer.s

* Remove asm/input_buffer.o from ldscript.txt

* Minor cleanup in sprite_2.c

* gUnknown_080D7F54 -> sPalAeroEggHit

* Rename AeroEggResetPos back to AeroEggMove

+++ include/constants/zones.h
@@ -55,7 +55,7 @@
 #define IS_FINAL_OR_EXTRA_STAGE(lvl) ((IS_FINAL_STAGE(lvl)) || (IS_EXTRA_STAGE(lvl)))
 
 #define ZONE_TIME_TO_INT(minutes, seconds)                                              \
-    (((minutes * 60) + seconds) * GBA_FRAMES_PER_SECOND)
+    (int)(((minutes * 60.) + seconds) * GBA_FRAMES_PER_SECOND)
 #define MAX_COURSE_TIME (ZONE_TIME_TO_INT(10, 0))
 
 #endif // GUARD_ZONES_H
